# express-prisma-swagger/src/modules/module1/controllers/ping.ts
import { Post, Get, Route } from "tsoa";

interface PingResponse {
  message: string;
}

@Route("resrse")
export default class PingController {
  @Get("/")
  public async getMessage(): Promise<PingResponse> {
    return {
      message: "hello",
    };
  }

  @Post("/")
  public async postMessage(): Promise<PingResponse> {
    return {
      message: "post",
    };
  }
}