# express-typeorm-swagger/src/entities/user.entity.ts
import { Entity, Column, PrimaryGeneratedColumn } from "typeorm"

@Entity()
export class User {
    @PrimaryGeneratedColumn()
    id!: number

    @Column()
    firstName?: string

    @Column()
    lastName?: string
}